requirements.txt
google-cloud-storage
pymongo